refactor: extract fs utilities
