+++ packages/util/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./fs";
 export * from "./git";
 export * from "./introspection";
 export * from "./metadata";
